pyproject.toml
[project]
name = "nl2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "en-core-web-sm",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "networkx>=3.6.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rapidfuzz>=3.14.3",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.48",
    "sqlglot>=29.0.1",
    "uvicorn>=0.41.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
